fix/ Style page that display secret password
